Update odd no's between two intervals.cpp

+++ Beginner/odd no's between two intervals.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-	int N,Q,i;
+	long int N,Q,i;
 	cin>>N>>Q;
 	for(i=N+1;i<Q;i++)
 	if(i%2==1)
